params_reader package renamed to paramsreader (to match convention)
