+++ src/main/java/com/blackhearth/blockchain/paramsreader/Pair.java
@@ -1,4 +1,4 @@
-package com.blackhearth.blockchain.params_reader;
+package com.blackhearth.blockchain.paramsreader;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
