+++ src/main/java/com/blackhearth/blockchain/paramsreader/ParamService.java
@@ -1,4 +1,4 @@
-package com.blackhearth.blockchain.params_reader;
+package com.blackhearth.blockchain.paramsreader;
 
 public interface ParamService {
     Pair<RunParams, Object> getValue(RunParams param, String[] args, int paramNumber);
